reorg and comments
